placing balls
