--- Controller.java
@@ -1,5 +1,0 @@
-package sample;
-
-public class Controller {
-
-}
