+++ src/sample/Controller.java
@@ -1,4 +1,5 @@
 package sample;
 
 public class Controller {
+
 }
